+++ src/actors/builtin/storage_miner/storage_miner_actor.id
@@ -1,7 +1,6 @@
 import abi "github.com/filecoin-project/specs/actors/abi"
 import autil "github.com/filecoin-project/specs/actors/util"
 import address "github.com/filecoin-project/go-address"
-import filcrypto "github.com/filecoin-project/specs/algorithms/crypto"
 import peer "github.com/libp2p/go-libp2p-core/peer"
 
 type MinerPoStState union {
@@ -105,7 +104,7 @@ type MinerInfo struct {
     // sign messages sent on behalf of this miner to commit sectors, submit PoSts, and
     // other day to day miner activities.
     Worker                  address.Address  // Must be an ID-address.
-    WorkerVRFKey            filcrypto.VRFPublicKey
+    WorkerVRFKey            address.Address  // Must be a SECP or BLS address
 
     // Libp2p identity that should be used when connecting to this miner.
     PeerId                  peer.ID
