+++ src/algorithms/crypto/randomness.md
@@ -59,5 +59,5 @@ buffer.append(other needed serialized inputs)
 randomness = SHA256(buffer)
 ```
 
-{{< readfile file="randomness.go" code="true" lang="go" >}}
+{{< readfile file="/docs/actors/crypto/randomness.go" code="true" lang="go" >}}
 {{< readfile file="chain.go" code="true" lang="go" >}}
