+++ src/systems/filecoin_blockchain/storage_power_consensus/storage_power_consensus_subsystem.go
@@ -7,6 +7,7 @@ import (
 	abi "github.com/filecoin-project/specs/actors/abi"
 	builtin "github.com/filecoin-project/specs/actors/builtin"
 	spowact "github.com/filecoin-project/specs/actors/builtin/storage_power"
+	acrypto "github.com/filecoin-project/specs/actors/crypto"
 	inds "github.com/filecoin-project/specs/actors/runtime/indices"
 	filcrypto "github.com/filecoin-project/specs/algorithms/crypto"
 	block "github.com/filecoin-project/specs/systems/filecoin_blockchain/struct/block"
@@ -37,7 +38,7 @@ func (spc *StoragePowerConsensusSubsystem_I) ComputeChainWeight(tipset chain.Tip
 func (spc *StoragePowerConsensusSubsystem_I) IsWinningPartialTicket(stateTree stateTree.StateTree, inds inds.Indices, partialTicket abi.PartialTicket, sectorUtilization abi.StoragePower, numSectors util.UVarint) bool {
 
 	// finalize the partial ticket
-	challengeTicket := filcrypto.SHA256(partialTicket)
+	challengeTicket := acrypto.SHA256(abi.Bytes(partialTicket))
 
 	networkPower := inds.TotalNetworkEffectivePower()
