+++ src/systems/filecoin_blockchain/struct/block/election.go
@@ -3,6 +3,7 @@ package block
 import (
 	addr "github.com/filecoin-project/go-address"
 	abi "github.com/filecoin-project/specs/actors/abi"
+	acrypto "github.com/filecoin-project/specs/actors/crypto"
 	filcrypto "github.com/filecoin-project/specs/algorithms/crypto"
 	util "github.com/filecoin-project/specs/util"
 )
@@ -12,11 +13,10 @@ func (tix *Ticket_I) ValidateSyntax() bool {
 }
 
 func (tix *Ticket_I) Verify(randomness util.Bytes, pk filcrypto.VRFPublicKey, minerActorAddr addr.Address) bool {
-
-	input := filcrypto.DeriveRandWithMinerAddr(filcrypto.DomainSeparationTag_TicketProduction, randomness, minerActorAddr)
+	input := acrypto.DeriveRandWithMinerAddr(acrypto.DomainSeparationTag_TicketProduction, randomness, minerActorAddr)
 	return tix.VRFResult_.Verify(input, pk)
 }
 
 func (tix *Ticket_I) DrawRandomness(epoch abi.ChainEpoch) util.Bytes {
-	return filcrypto.DeriveRandWithEpoch(filcrypto.DomainSeparationTag_TicketDrawing, tix.Output(), int(epoch))
+	return acrypto.DeriveRandWithEpoch(acrypto.DomainSeparationTag_TicketDrawing, tix.Output(), int(epoch))
 }
